requirements.txt
# Core dependencies for video generation
numpy>=1.21.0
scipy>=1.7.0
h5py>=3.7.0
tqdm>=4.62.0

# Video processing
scikit-video>=1.1.11
opencv-python>=4.5.0

# MuJoCo and MyoSuite
mujoco>=2.3.0
myosuite>=1.0.0

# Google Cloud Storage support
google-cloud-storage>=2.0.0

# Optional: For better performance and additional features
# matplotlib>=3.5.0  # For plotting (if needed)
# pandas>=1.3.0      # For data manipulation (if needed)
# jupyter>=1.0.0     # For Jupyter notebooks (if needed)

# Development and testing (optional)
# pytest>=6.0.0      # For testing
# black>=21.0.0      # For code formatting
# flake8>=3.9.0      # For linting
